requirements.txt
python-telegram-bot
notion-client
cryptography
fastapi
uvicorn